:recycle: [refactory]: New directory(Header)
- Add Header component
- Collections.tsx component refactory to Header
