+++ src/pages/header/page/index.ts
@@ -1,0 +1,1 @@
+export { default as Header } from './Header';
